move getUserWithPermissions into testutils as is used exclusively by tests. Update function content so that all updates happen via the endpoints to ensure all listeners/validation etc still occur
